[jit] separate QualifiedName into its own file
We want to use this to generically represent qualified names for things
(e.g. ScriptClasses). So separating out the generic stuff from the
python_print-specific stuff and sticking it in its own file.

Test Plan:
No behavioral changes, so existing test_jit.py tests
