+++ torch/csrc/jit/passes/python_print.cpp
@@ -4,6 +4,7 @@
 #include <torch/csrc/jit/ir.h>
 #include <torch/csrc/jit/ir_views.h>
 #include <torch/csrc/jit/passes/python_print.h>
+#include <torch/csrc/jit/qualified_name.h>
 #include <torch/csrc/jit/resource_guard.h>
 #include <torch/csrc/jit/script/error_report.h>
 #include <torch/csrc/jit/script/module.h>
@@ -84,48 +85,33 @@ static bool isValidIdentifier(const std::string& name) {
   return true;
 }
 
-// handles names of the form, e.g., self.a.b
-// if a field is not a valid identifier, then it will print as, e.g.
-// getattr(self, "0").b
-struct QualifiedName;
-using QualifiedNamePtr = c10::intrusive_ptr<QualifiedName>;
-struct QualifiedName : c10::intrusive_ptr_target {
-  QualifiedName(QualifiedNamePtr prefix, std::string name)
-      : prefix_(std::move(prefix)), name_(std::move(name)) {}
-  QualifiedNamePtr prefix_;
-  std::string name_;
-  static QualifiedNamePtr create(QualifiedNamePtr prefix, std::string name) {
-    return c10::make_intrusive<QualifiedName>(
-        std::move(prefix), std::move(name));
-  }
-  static QualifiedNamePtr create(std::string name) {
-    return c10::make_intrusive<QualifiedName>(
-        QualifiedNamePtr(), std::move(name));
-  }
-  std::string str() const {
-    std::stringstream ss;
-    emit(ss);
-    return ss.str();
-  }
-
- private:
-  void emit(std::ostream& out) const {
-    if (isValidIdentifier(name_)) {
-      if (prefix_) {
-        prefix_->emit(out);
-        out << ".";
-      }
-      out << name_;
-    } else {
-      AT_ASSERT(prefix_);
-      out << "getattr(";
-      prefix_->emit(out);
-      out << ", ";
-      printQuotedString(out, name_);
-      out << ")";
+static void emitQualifiedName(std::ostream& out, QualifiedNamePtr name) {
+  const auto& name_ = name->name_;
+  const auto& prefix_ = name->prefix_;
+  if (isValidIdentifier(name_)) {
+    if (prefix_) {
+      emitQualifiedName(out, prefix_);
+      out << ".";
     }
+    out << name_;
+  } else {
+    AT_ASSERT(prefix_);
+    out << "getattr(";
+    emitQualifiedName(out, prefix_);
+    out << ", ";
+    printQuotedString(out, name_);
+    out << ")";
   }
-};
+}
+
+// Get a stringified version of the qualified name.
+// if a field is not a valid Python identifier, then it will print as, e.g.
+// getattr(self, "0").b
+static std::string getValidQualifiedName(QualifiedNamePtr name) {
+  std::stringstream ss;
+  emitQualifiedName(ss, std::move(name));
+  return ss.str();
+}
 
 void createTensorToParameterNameMap(
     const script::Module& module,
@@ -1143,7 +1129,7 @@ struct PythonPrintPass {
           extra_ivalue_names) {
     std::vector<std::string> ivalue_names =
         fmap(method.initial_ivalues(), [&](const script::Slot& slot) {
-          return extra_ivalue_names.at(slot)->str();
+          return getValidQualifiedName(extra_ivalue_names.at(slot));
         });
     const std::string& name = method.name();
     Graph& graph = *method.graph();
